#2: Érmehamisítás finomítása

+++ 02Nap/02SikidomokTerulete/IPlane.cs
@@ -4,6 +4,6 @@
     {
         //a public kulcsszóra itt nincs szükség, mert ami a felületen látszik, az publikus
         //aminek nem kell látszania, az nem látszik
-        int Area();
+        double Area();
     }
 }
